Fix flashlist to support percentage paddings
